framework: Split diagram_state files out of diagram_context
