+++ systems/framework/BUILD.bazel
@@ -30,6 +30,7 @@ drake_cc_package_library(
         ":diagram_continuous_state",
         ":diagram_discrete_values",
         ":diagram_output_port",
+        ":diagram_state",
         ":discrete_values",
         ":event_collection",
         ":framework_common",
@@ -552,13 +553,31 @@ drake_cc_library(
 )
 
 drake_cc_library(
+    name = "diagram_state",
+    srcs = ["diagram_state.cc"],
+    hdrs = ["diagram_state.h"],
+    deps = [
+        ":abstract_values",
+        ":continuous_state",
+        ":diagram_continuous_state",
+        ":diagram_discrete_values",
+        ":discrete_values",
+        ":state",
+        "//common:default_scalars",
+        "//common:essential",
+    ],
+)
+
+drake_cc_library(
     name = "diagram_context",
     srcs = ["diagram_context.cc"],
     hdrs = ["diagram_context.h"],
     deps = [
         ":context",
-        ":diagram_continuous_state",
-        ":diagram_discrete_values",
+        ":diagram_state",
+        ":framework_common",
+        ":parameters",
+        ":state",
         ":vector",
         "//common:default_scalars",
         "//common:essential",
